--- src/main/java/dev/rubikon/api/commons/Event.java
@@ -1,4 +1,0 @@
-package dev.rubikon.api.commons;
-
-public interface Event {
-}
