--- src/main/java/dev/rubikon/api/commons/Serializable.java
@@ -1,12 +1,0 @@
-package dev.rubikon.api.commons;
-
-import com.google.gson.JsonObject;
-
-/**
- * This interface is used to save and load data from the config(json) file.
- */
-public interface Serializable {
-
-    JsonObject save();
-    void load(JsonObject jsonObject);
-}
