--- src/main/java/dev/rubikon/api/commons/Toggleable.java
@@ -1,11 +1,0 @@
-package dev.rubikon.api.commons;
-
-
-/**
- * This interface is used to mark features that can be toggled on and off.
- */
-public interface Toggleable {
-    void toggle();
-    void onEnable();
-    void onDisable();
-}
