+++ src/main/java/dev/rubikon/commons/Event.java
@@ -1,0 +1,4 @@
+package dev.rubikon.commons;
+
+public interface Event {
+}
