+++ src/main/java/dev/rubikon/commons/Store.java
@@ -1,0 +1,11 @@
+package dev.rubikon.commons;
+
+import java.util.Collection;
+
+public interface Store<T, U> {
+    default void add(U item) {};
+    default void add(T name, U item) {};
+    U find(T name);
+    Collection<T> names();
+    Collection<U> all();
+}
