+++ src/main/java/dev/rubikon/commons/Toggleable.java
@@ -1,0 +1,11 @@
+package dev.rubikon.commons;
+
+
+/**
+ * This interface is used to mark features that can be toggled on and off.
+ */
+public interface Toggleable {
+    void toggle();
+    default void onEnable() {};
+    default void onDisable() {};
+}
