+++ src/main/java/dev/rubikon/events/ScreenRenderEvent.java
@@ -1,6 +1,6 @@
 package dev.rubikon.events;
 
-import dev.rubikon.api.commons.Event;
+import dev.rubikon.commons.Event;
 
 /**
  * @param delta The time in seconds since the last frame.
