+++ src/main/java/dev/rubikon/events/TickEvent.java
@@ -1,6 +1,6 @@
 package dev.rubikon.events;
 
-import dev.rubikon.api.commons.Event;
+import dev.rubikon.commons.Event;
 
 /**
  * Invoked every game tick.
