+++ src/main/java/dev/rubikon/mixin/MinecraftClientMixin.java
@@ -2,9 +2,7 @@ package dev.rubikon.mixin;
 
 import dev.rubikon.Rubikon;
 import dev.rubikon.events.TickEvent;
-import dev.rubikon.renderer.core.imgui.ImGuiContext;
-import dev.rubikon.renderer.core.nanovg.NVContext;
-import dev.rubikon.stores.Stores;
+import dev.rubikon.renderer.core.Renderer;
 import net.minecraft.client.MinecraftClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -15,14 +13,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class MinecraftClientMixin {
 	@Inject(at = @At("TAIL"), method = "<init>")
 	private void init(CallbackInfo info) {
-		//only for initializing
-		NVContext.initialize();
-		ImGuiContext.initialize();
-		Stores.init();
+		// We need to initialize the renderer here because the game crashes if we do it in Rubikon#onInitialize()
+		Renderer.getInstance().init();
 	}
 	@Inject(at = @At("TAIL"), method = "tick")
 	private void tick(CallbackInfo info) {
 		Rubikon.getEventPubSub().publish(new TickEvent());
 	}
-
 }
