+++ src/main/java/dev/rubikon/things/features/misc/Logo.java
@@ -1,0 +1,58 @@
+package dev.rubikon.things.features.misc;
+
+import dev.rubikon.events.ScreenRenderEvent;
+import dev.rubikon.renderer.core.Renderer;
+import dev.rubikon.renderer.core.nanovg.NVContext;
+import dev.rubikon.things.features.Feature;
+import io.github.nevalackin.radbus.Listen;
+import org.lwjgl.glfw.GLFW;
+import org.lwjgl.nanovg.NVGColor;
+import org.lwjgl.nanovg.NVGPaint;
+import org.lwjgl.system.MemoryStack;
+
+import java.nio.IntBuffer;
+
+import static org.lwjgl.nanovg.NanoVG.*;
+
+public class Logo extends Feature {
+    public Logo() {
+        super("Logo","Lets you toggle rendering of the client logo", GLFW.GLFW_KEY_L);
+    }
+
+    @Override
+    public void onEnable() {
+
+    }
+
+    @Listen
+    public void onRender(ScreenRenderEvent event) {
+        int image = Renderer.getInstance().find("rubikon-icon");
+
+        //render the text
+        NVContext.draw(ctx -> {
+            //alloc memory
+            //TODO implement rasterizer
+            try (MemoryStack stack = MemoryStack.stackPush()) {
+                //get default image size
+                IntBuffer pW = stack.mallocInt(1);
+                IntBuffer pH = stack.mallocInt(1);
+                //set it
+                nvgImageSize(ctx, image, pW, pH);
+                //clear paths
+                nvgBeginPath(ctx);
+                nvgRect(ctx, 2, 4, 25, 25);
+                NVGPaint logoPaint = NVGPaint.create();
+                //creates fbo as a image and binds it onto the paint
+                nvgImagePattern(ctx, 0.5f, 1, 30, 30, 0.f, image, 1.f, logoPaint);
+                nvgFillPaint(ctx, logoPaint);
+                nvgFill(ctx);
+            }
+            nvgFontFace(ctx,"sfui-bold");
+            nvgFontSize(ctx,24.f);
+            NVGColor fillcolor = NVContext.nvgColor(-1);
+            nvgFillColor(ctx,fillcolor);
+            //draw text
+            nvgText(ctx,25,25,"Rubikon");
+        });
+    }
+}
